Refactor: 회의실 세팅 사이드바 컴포넌트 분리 및 active 상태 반영
